+++ src/test/java/fixture/BettingFixture.java
@@ -1,9 +1,10 @@
 package fixture;
 
 import blackjack.gamer.Betting;
+import java.math.BigDecimal;
 
 public class BettingFixture {
-    public static Betting createBetting(Integer bettingAmount) {
-        return new Betting(bettingAmount);
+    public static Betting createBetting(int bettingAmount) {
+        return new Betting(BigDecimal.valueOf(bettingAmount));
     }
 }
